reworked OrderDaoDb and ProductDaoDb so now both working with one table of products
